refactor(Application): change the parameters of validation methods from string to list
validation the input string by converting String to List in the method getPlayerNumbers.
In method validateStringThreeNumberDuplicate, Each of the validation methods called in this function handles exceptions.
In method validateThreeNaturalNumber, Change the way to determine if the number is in the range to use stream.
In method validateThreeNumberDuplicate, Change the way to determine if three numbers are duplicated using stream.